+++ src/lib.rs
@@ -4,6 +4,9 @@ pub use activities_api::ActivitiesApi;
 mod configuration;
 pub use configuration::Configuration;
 
+mod error;
+pub use error::Error;
+
 mod api_client;
 pub use api_client::ApiClient;